+++ modules/tinymce/src/plugins/lists/main/ts/actions/ToggleList.ts
@@ -11,7 +11,7 @@ import * as Bookmark from '../core/Bookmark';
 import { listToggleActionFromListName } from '../core/ListAction';
 import * as NodeType from '../core/NodeType';
 import * as Selection from '../core/Selection';
-import { isCustomList, isWithinNonEditableList } from '../core/Util';
+import { hasNonEditableBlocksSelected, isCustomList, isWithinNonEditableList } from '../core/Util';
 import { flattenListSelection } from './Indendation';
 
 interface ListDetail {
@@ -344,7 +344,7 @@ const toggleSingleList = (editor: Editor, parentList: HTMLElement | null, listNa
 
 const toggleList = (editor: Editor, listName: 'UL' | 'OL' | 'DL', _detail: ListDetail | null): void => {
   const parentList = Selection.getParentList(editor);
-  if (isWithinNonEditableList(editor, parentList)) {
+  if (isWithinNonEditableList(editor, parentList) || hasNonEditableBlocksSelected(editor)) {
     return;
   }